LPS-189437 Rename classes to follow pattern "public class XyzAbc extends BaseAbc"
