+++ modules/apps/fragment/fragment-service/src/main/java/com/liferay/fragment/internal/exportimport/content/processor/CollectionSelectorEditableValuesConfigurationExportImportContentProcessor.java
@@ -1,0 +1,114 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.fragment.internal.exportimport.content.processor;
+
+import com.liferay.asset.list.model.AssetListEntry;
+import com.liferay.asset.list.service.AssetListEntryLocalService;
+import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
+import com.liferay.exportimport.kernel.lar.PortletDataContext;
+import com.liferay.exportimport.kernel.lar.StagedModelDataHandlerUtil;
+import com.liferay.fragment.util.configuration.FragmentEntryConfigurationParser;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.model.StagedModel;
+import com.liferay.portal.kernel.xml.Element;
+
+import java.util.Map;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Víctor Galán
+ */
+@Component(
+	property = "content.processor.type=FragmentEntryLinkEditableValues",
+	service = ExportImportContentProcessor.class
+)
+public class
+	CollectionSelectorEditableValuesConfigurationExportImportContentProcessor
+		extends BaseEditableValuesConfigurationExportImportContentProcessor {
+
+	@Override
+	protected String getConfigurationType() {
+		return "collectionSelector";
+	}
+
+	@Override
+	protected FragmentEntryConfigurationParser
+		getFragmentEntryConfigurationParser() {
+
+		return _fragmentEntryConfigurationParser;
+	}
+
+	@Override
+	protected void replaceExportContentReferences(
+			PortletDataContext portletDataContext, StagedModel stagedModel,
+			JSONObject configurationValueJSONObject,
+			boolean exportReferencedContent)
+		throws Exception {
+
+		if (!configurationValueJSONObject.has("classPK")) {
+			return;
+		}
+
+		AssetListEntry assetListEntry =
+			_assetListEntryLocalService.fetchAssetListEntry(
+				configurationValueJSONObject.getLong("classPK"));
+
+		if (assetListEntry == null) {
+			return;
+		}
+
+		if (exportReferencedContent) {
+			StagedModelDataHandlerUtil.exportReferenceStagedModel(
+				portletDataContext, assetListEntry, stagedModel,
+				PortletDataContext.REFERENCE_TYPE_DEPENDENCY);
+		}
+		else {
+			Element entityElement = portletDataContext.getExportDataElement(
+				assetListEntry);
+
+			portletDataContext.addReferenceElement(
+				assetListEntry, entityElement, stagedModel,
+				PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+		}
+	}
+
+	@Override
+	protected void replaceImportContentReferences(
+		PortletDataContext portletDataContext,
+		JSONObject configurationValueJSONObject) {
+
+		if (!configurationValueJSONObject.has("classPK")) {
+			return;
+		}
+
+		Map<Long, Long> assetListEntryNewPrimaryKeys =
+			(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
+				AssetListEntry.class.getName());
+
+		configurationValueJSONObject.put(
+			"classPK",
+			assetListEntryNewPrimaryKeys.getOrDefault(
+				configurationValueJSONObject.getLong("classPK"), 0L));
+	}
+
+	@Reference
+	private AssetListEntryLocalService _assetListEntryLocalService;
+
+	@Reference
+	private FragmentEntryConfigurationParser _fragmentEntryConfigurationParser;
+
+}
