+++ modules/apps/fragment/fragment-service/src/main/java/com/liferay/fragment/internal/exportimport/content/processor/NavigationMenuSelectorEditableValuesConfigurationExportImportContentProcessor.java
@@ -14,18 +14,18 @@
 
 package com.liferay.fragment.internal.exportimport.content.processor;
 
-import com.liferay.asset.kernel.model.AssetCategory;
-import com.liferay.asset.kernel.model.AssetVocabulary;
-import com.liferay.asset.kernel.service.AssetCategoryLocalService;
-import com.liferay.asset.kernel.service.AssetVocabularyLocalService;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.kernel.lar.StagedModelDataHandlerUtil;
 import com.liferay.fragment.util.configuration.FragmentEntryConfigurationParser;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.StagedModel;
-import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.xml.Element;
+import com.liferay.site.navigation.model.SiteNavigationMenu;
+import com.liferay.site.navigation.model.SiteNavigationMenuItem;
+import com.liferay.site.navigation.service.SiteNavigationMenuLocalService;
 
 import java.util.Map;
 
@@ -39,12 +39,13 @@ import org.osgi.service.component.annotations.Reference;
 	property = "content.processor.type=FragmentEntryLinkEditableValues",
 	service = ExportImportContentProcessor.class
 )
-public class EditableValuesCategoryTreeNodeSelectorExportImportContentProcessor
-	extends BaseEditableValuesConfigurationExportImportContentProcessor {
+public class
+	NavigationMenuSelectorEditableValuesConfigurationExportImportContentProcessor
+		extends BaseEditableValuesConfigurationExportImportContentProcessor {
 
 	@Override
 	protected String getConfigurationType() {
-		return "categoryTreeNodeSelector";
+		return "navigationMenuSelector";
 	}
 
 	@Override
@@ -62,25 +63,20 @@ public class EditableValuesCategoryTreeNodeSelectorExportImportContentProcessor
 			boolean exportReferencedContent)
 		throws Exception {
 
-		long assetCategoryTreeNodeId = GetterUtil.getLong(
-			configurationValueJSONObject.getString("categoryTreeNodeId"));
-
-		if (assetCategoryTreeNodeId == 0) {
-			return;
-		}
+		long siteNavigationMenuId = configurationValueJSONObject.getLong(
+			"siteNavigationMenuId");
 
 		StagedModel stagedModel = null;
 
-		String assetCategoryTreeNodeType =
-			configurationValueJSONObject.getString("categoryTreeNodeType");
-
-		if (assetCategoryTreeNodeType.equals("Vocabulary")) {
-			stagedModel = _assetVocabularyLocalService.fetchAssetVocabulary(
-				assetCategoryTreeNodeId);
+		if (siteNavigationMenuId > 0) {
+			stagedModel =
+				_siteNavigationMenuLocalService.fetchSiteNavigationMenu(
+					siteNavigationMenuId);
 		}
-		else if (assetCategoryTreeNodeType.equals("Category")) {
-			stagedModel = _assetCategoryLocalService.fetchAssetCategory(
-				assetCategoryTreeNodeId);
+		else {
+			stagedModel = _layoutLocalService.fetchLayout(
+				configurationValueJSONObject.getLong(
+					"parentSiteNavigationMenuItemId"));
 		}
 
 		if (stagedModel == null) {
@@ -107,45 +103,51 @@ public class EditableValuesCategoryTreeNodeSelectorExportImportContentProcessor
 		PortletDataContext portletDataContext,
 		JSONObject configurationValueJSONObject) {
 
-		long assetCategoryTreeNodeId = GetterUtil.getLong(
-			configurationValueJSONObject.getString("categoryTreeNodeId"));
+		long siteNavigationMenuId = configurationValueJSONObject.getLong(
+			"siteNavigationMenuId");
 
-		if (assetCategoryTreeNodeId == 0) {
-			return;
-		}
+		long parentSiteNavigationMenuItemId =
+			configurationValueJSONObject.getLong(
+				"parentSiteNavigationMenuItemId");
 
-		String assetCategoryTreeNodeType =
-			configurationValueJSONObject.getString("categoryTreeNodeType");
-
-		if (assetCategoryTreeNodeType.equals("Vocabulary")) {
-			Map<Long, Long> assetVocabularyNewPrimaryKeys =
+		if (siteNavigationMenuId == 0) {
+			Map<Long, Long> layoutNewPrimaryKeys =
 				(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
-					AssetVocabulary.class.getName());
+					Layout.class.getName());
 
 			configurationValueJSONObject.put(
-				"categoryTreeNodeId",
-				assetVocabularyNewPrimaryKeys.getOrDefault(
-					assetCategoryTreeNodeId, 0L));
+				"parentSiteNavigationMenuItemId",
+				layoutNewPrimaryKeys.getOrDefault(
+					parentSiteNavigationMenuItemId, 0L));
 		}
-		else if (assetCategoryTreeNodeType.equals("Category")) {
-			Map<Long, Long> assetVocabularyNewPrimaryKeys =
+		else {
+			Map<Long, Long> siteNavigationMenuNewPrimaryKeys =
 				(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
-					AssetCategory.class.getName());
+					SiteNavigationMenu.class.getName());
 
 			configurationValueJSONObject.put(
-				"categoryTreeNodeId",
-				assetVocabularyNewPrimaryKeys.getOrDefault(
-					assetCategoryTreeNodeId, 0L));
+				"siteNavigationMenuId",
+				siteNavigationMenuNewPrimaryKeys.getOrDefault(
+					siteNavigationMenuId, 0L));
+
+			Map<Long, Long> siteNavigationMenuItemNewPrimaryKeys =
+				(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
+					SiteNavigationMenuItem.class.getName());
+
+			configurationValueJSONObject.put(
+				"parentSiteNavigationMenuItemId",
+				siteNavigationMenuItemNewPrimaryKeys.getOrDefault(
+					parentSiteNavigationMenuItemId, 0L));
 		}
 	}
 
 	@Reference
-	private AssetCategoryLocalService _assetCategoryLocalService;
-
-	@Reference
-	private AssetVocabularyLocalService _assetVocabularyLocalService;
-
-	@Reference
 	private FragmentEntryConfigurationParser _fragmentEntryConfigurationParser;
+
+	@Reference
+	private LayoutLocalService _layoutLocalService;
+
+	@Reference
+	private SiteNavigationMenuLocalService _siteNavigationMenuLocalService;
 
 }
