+++ modules/apps/fragment/fragment-service/src/main/java/com/liferay/fragment/internal/exportimport/content/processor/URLEditableValuesConfigurationExportImportContentProcessor.java
@@ -14,14 +14,15 @@
 
 package com.liferay.fragment.internal.exportimport.content.processor;
 
-import com.liferay.asset.list.model.AssetListEntry;
-import com.liferay.asset.list.service.AssetListEntryLocalService;
 import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
 import com.liferay.exportimport.kernel.lar.StagedModelDataHandlerUtil;
 import com.liferay.fragment.util.configuration.FragmentEntryConfigurationParser;
 import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.StagedModel;
+import com.liferay.portal.kernel.service.LayoutLocalService;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.xml.Element;
 
 import java.util.Map;
@@ -30,18 +31,18 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Víctor Galán
+ * @author Ricardo Couso
  */
 @Component(
 	property = "content.processor.type=FragmentEntryLinkEditableValues",
 	service = ExportImportContentProcessor.class
 )
-public class EditableValuesCollectionSelectorExportImportContentProcessor
+public class URLEditableValuesConfigurationExportImportContentProcessor
 	extends BaseEditableValuesConfigurationExportImportContentProcessor {
 
 	@Override
 	protected String getConfigurationType() {
-		return "collectionSelector";
+		return "url";
 	}
 
 	@Override
@@ -58,30 +59,13 @@ public class EditableValuesCollectionSelectorExportImportContentProcessor
 			boolean exportReferencedContent)
 		throws Exception {
 
-		if (!configurationValueJSONObject.has("classPK")) {
-			return;
-		}
+		if ((configurationValueJSONObject != null) &&
+			configurationValueJSONObject.has("layout")) {
 
-		AssetListEntry assetListEntry =
-			_assetListEntryLocalService.fetchAssetListEntry(
-				configurationValueJSONObject.getLong("classPK"));
-
-		if (assetListEntry == null) {
-			return;
-		}
-
-		if (exportReferencedContent) {
-			StagedModelDataHandlerUtil.exportReferenceStagedModel(
-				portletDataContext, assetListEntry, stagedModel,
-				PortletDataContext.REFERENCE_TYPE_DEPENDENCY);
-		}
-		else {
-			Element entityElement = portletDataContext.getExportDataElement(
-				assetListEntry);
-
-			portletDataContext.addReferenceElement(
-				assetListEntry, entityElement, stagedModel,
-				PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+			_exportLayoutReferences(
+				portletDataContext, stagedModel,
+				configurationValueJSONObject.getJSONObject("layout"),
+				exportReferencedContent);
 		}
 	}
 
@@ -90,24 +74,86 @@ public class EditableValuesCollectionSelectorExportImportContentProcessor
 		PortletDataContext portletDataContext,
 		JSONObject configurationValueJSONObject) {
 
-		if (!configurationValueJSONObject.has("classPK")) {
+		if ((configurationValueJSONObject != null) &&
+			configurationValueJSONObject.has("layout")) {
+
+			_replaceImportLayoutReferences(
+				configurationValueJSONObject.getJSONObject("layout"),
+				portletDataContext);
+		}
+	}
+
+	private void _exportLayoutReferences(
+			PortletDataContext portletDataContext,
+			StagedModel referrerStagedModel, JSONObject layoutJSONObject,
+			boolean exportReferencedContent)
+		throws Exception {
+
+		if (layoutJSONObject.length() == 0) {
 			return;
 		}
 
-		Map<Long, Long> assetListEntryNewPrimaryKeys =
-			(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
-				AssetListEntry.class.getName());
+		Layout layout = _layoutLocalService.fetchLayout(
+			layoutJSONObject.getLong("groupId"),
+			layoutJSONObject.getBoolean("privateLayout"),
+			layoutJSONObject.getLong("layoutId"));
 
-		configurationValueJSONObject.put(
-			"classPK",
-			assetListEntryNewPrimaryKeys.getOrDefault(
-				configurationValueJSONObject.getLong("classPK"), 0L));
+		if (layout == null) {
+			return;
+		}
+
+		layoutJSONObject.put("plid", layout.getPlid());
+
+		if (exportReferencedContent) {
+			StagedModelDataHandlerUtil.exportReferenceStagedModel(
+				portletDataContext, referrerStagedModel, layout,
+				PortletDataContext.REFERENCE_TYPE_DEPENDENCY);
+		}
+		else {
+			Element entityElement = portletDataContext.getExportDataElement(
+				referrerStagedModel);
+
+			portletDataContext.addReferenceElement(
+				referrerStagedModel, entityElement, layout,
+				PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
+		}
+	}
+
+	private void _replaceImportLayoutReferences(
+		JSONObject layoutJSONObject, PortletDataContext portletDataContext) {
+
+		if (layoutJSONObject.length() == 0) {
+			return;
+		}
+
+		long plid = GetterUtil.getLong(layoutJSONObject.remove("plid"));
+
+		Map<Long, Long> layoutNewPrimaryKeys =
+			(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
+				Layout.class.getName());
+
+		Layout layout = _layoutLocalService.fetchLayout(
+			layoutNewPrimaryKeys.getOrDefault(plid, 0L));
+
+		if (layout == null) {
+			return;
+		}
+
+		layoutJSONObject.put(
+			"groupId", layout.getGroupId()
+		).put(
+			"layoutId", layout.getLayoutId()
+		).put(
+			"layoutUuid", layout.getUuid()
+		).put(
+			"privateLayout", layout.isPrivateLayout()
+		);
 	}
 
 	@Reference
-	private AssetListEntryLocalService _assetListEntryLocalService;
+	private FragmentEntryConfigurationParser _fragmentEntryConfigurationParser;
 
 	@Reference
-	private FragmentEntryConfigurationParser _fragmentEntryConfigurationParser;
+	private LayoutLocalService _layoutLocalService;
 
 }
